# api-definition/src/main/resources/static/V1/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Task API
  description: The purpose of this API is for external systems to manage tasks.
servers:
  - url: http://localhost:8080
    description: Task API location
...
paths:
  /projects:
    get:
      description: Returns all the projects from the system that the user has access to
      tags:
        - "Projects"
      operationId: getProjects
      x-paginated-object:
        - gr.rk.tasks.V1.dto.ProjectDTO
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
        - in: query
          name: identifier
          required: false
          schema:
            type: string
        - in: query
          name: name
          required: false
          schema:
            type: string
        - in: query
          name: creationDateFrom
          required: false
          schema:
            type: string
            pattern: '^([0-9]+)-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])[Tt]([01][0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]|60)(\.[0-9]+)?(([Zz])|([\+|\-]([01][0-9]|2[0-3]):[0-5][0-9]))$'
        - in: query
          name: creationDateTo
          required: false
          schema:
            type: string
            pattern: '^([0-9]+)-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])[Tt]([01][0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]|60)(\.[0-9]+)?(([Zz])|([\+|\-]([01][0-9]|2[0-3]):[0-5][0-9]))$'
        - in: query
          name: createdBy
          required: false
          schema:
            type: string
      responses:
        200:
          description: Paginated Projects
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                properties:
                  content:
                    items:
                      $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      description: Create a new Project
      tags:
        - "Projects"
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        "200":
          description: A Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /projects/{identifier}:
    get:
      description: Return specific project
      tags:
        - "Projects"
      operationId: getProject
      parameters:
        - name: identifier
          in: path
          description: identifier of the Project
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      description: Delete the project logical
      tags:
        - "Projects"
      operationId: deleteProject
      parameters:
      - name: identifier
        in: path
        description: identifier of the Project
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully logical delete.
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  schemas:
    Page:
      type: object
      properties:
        pageable:
          type: string
          example: INSTANCE
        last:
          type: boolean
        totalPages:
          type: integer
          example: 1
        totalElements:
          type: integer
          example: 5
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
        sort:
          type: object
          properties:
            sorted:
              type: boolean
              example: false
            unsorted:
              type: boolean
            empty:
                type: boolean
        first:
          type: boolean
        numberOfElements:
          type: integer
          example: 5
        empty:
          type: boolean
          example: false
    Project:
     type: object
     required:
       - name
       - createdBy
     properties:
       identifier:
         type: string
         description: Returned by GET, not used in POST/PUT/PATCH
         readOnly: true
       name:
         description: This is the name of the project
         type: string
       description:
         description: This is the description of the project
         type: string
       prefixIdentifier:
         description: The prefix part that every identifier haves. For example, if the prefix is <my-project> then the identifier will be <my-project-number>
         type: string
       creationDate:
         type: string
         pattern: '^([0-9]+)-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])[Tt]([01][0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]|60)(\.[0-9]+)?(([Zz])|([\+|\-]([01][0-9]|2[0-3]):[0-5][0-9]))$'
         description: Creation date and time
         example: "2021-01-30T08:30:00Z"
         readOnly: true
       createdBy:
         $ref: '#/components/schemas/User'
       realm:
         type: string
         description: This field is the realm from the IDP
         example: Application1
         readOnly: true
...
    Group:
      description: This is a group that a user belongs to
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        realm:
          type: string
          description: This field is the realm from the IDP
          example: Application1
          readOnly: true
        creationDate:
          type: string
          pattern: '^([0-9]+)-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])[Tt]([01][0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]|60)(\.[0-9]+)?(([Zz])|([\+|\-]([01][0-9]|2[0-3]):[0-5][0-9]))$'
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
          readOnly: true
        users:
          readOnly: true
          items:
            $ref: '#/components/schemas/User'
    User:
      description: This is a user object
      type: object
      required:
        - name
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        groups:
          readOnly: true
          items:
            $ref: "#/components/schemas/Group"
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum:
            - fail
            - error
          description: |
            The fail status indicating that there was a problem with the data submitted or some
            pre-condition on API call was not satisfied.
            The error status indicating that an error occurred in processing the request, i.e. Exception
            was thrown
        message:
          type: string
    Pageable:
      description: minimal Pageable query parameters
      type: object
      properties:
        page:
          type: integer
          description: Results page you want to retrieve (0..N)
        size:
          type: integer
          description: Number of records per page
        sort:
          description: Sorting criteria in the format -> property,asc|desc no space after comma!.
          example: id,asc
          type: string
  responses:
    ErrorResponse:
      description: Error payload
      content:
        application/json:
          schema:
            properties:
              errors:
                items:
                  $ref: '#/components/schemas/Error'